// src/models/pinPadType.ts
export enum PinPadType {
    openingApp = 'openingApp',
    backToApp = 'backToApp',
    pinChangeStage1 = 'pinChangeStage1',
    pinChangeStage2 = 'pinChangeStage2',
    firstConnexionStage1 = 'firstConnexionStage1',
    firstConnexionStage2 = 'firstConnexionStage2'
}

export enum PinPadTitle {
    openingApp = 'PINPAD.TITLE.OPENINGAPP',
    backToApp = 'PINPAD.TITLE.BACKTOAPP',
    pinChangeStage1 = 'PINPAD.TITLE.PINCHANGESTAGE1',
    pinChangeStage2 = 'PINPAD.TITLE.PINCHANGESTAGE2',
    firstConnexionStage1 = 'PINPAD.TITLE.FIRSTCONNEXIONSTAGE1',
    firstConnexionStage2 = 'PINPAD.TITLE.FIRSTCONNEXIONSTAGE2'
}